refactor: ProductCatogory를 ProductCategory 로 이름 변경
