+++ backend/sin/src/main/java/sin/sin/domain/productCategory/ProductCategory.java
@@ -1,4 +1,4 @@
-package sin.sin.domain.productCatogory;
+package sin.sin.domain.productCategory;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ProductCatogory {
+public class ProductCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
